% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsp.R
\name{hsp}
\alias{hsp}
\title{Hurdle Shifted Poisson model}
\usage{
hsp(data, h = 10, intervals = TRUE, level = 0.95, holdout = FALSE)
}
\arguments{
\item{data}{the vector of values (time sereis vector).}

\item{h}{forecasting horizon.}

\item{intervals}{binary, definign whether to construct prediction intervals or not.}

\item{level}{confidence level for prediction intervals.}

\item{holdout}{whether to use holdout of h observations or not.}
}
\value{
Function returns a model of a class "counter", which contains:
\itemize{
\item model - the name of the constructed model,
\item occurrence - ETS(A,N,N) model for demand intervals (from es() function),
\item sizes - ETS(A,N,N) model for demand sizes (from es() function),
\item fitted - fitted values,
\item forecast - forecasts of the function,
\item lower - lower bound of the prediction interval,
\item upper - upper bound of the prediction interval,
\item actuals - the provided actual values,
\item holdout - the actual values from the holdout (if holdout was set to TRUE),
\item level - confidence level used,
\item accuracy - the error measures for the data if the holdout was TRUE.
}
}
\description{
Construct HSP model on time series data
}
\details{
HSP model consists of two parts: Simple Exponential Smoothing for demand intervals
and the same method for the mean demand sizes (lambda parameter in Poisson distribution).
}
\examples{
y <- c(rpois(50,0.3),rpois(50,0.8))
test <- hsp(y)

}
\references{
\itemize{
\item Snyder, R. D., Ord, J. K., & Beaumont, A. (2012). Forecasting the
intermittent demand for slow-moving inventories: A modelling approach.
International Journal of Forecasting, 28(2), 485–496.
https://doi.org/10.1016/j.ijforecast.2011.03.009
}
}
\seealso{
\code{\link[counter]{negbin}, \link[smooth]{es}, \link[smooth]{iss}}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{models}
\keyword{ts}
