% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin.R
\name{negbin}
\alias{negbin}
\title{Dynamic Negative Binomial model}
\usage{
negbin(data, h = 10, intervals = TRUE, level = 0.95,
  holdout = FALSE)
}
\arguments{
\item{data}{the vector of values (time sereis vector).}

\item{h}{forecasting horizon.}

\item{intervals}{binary, definign whether to construct prediction intervals or not.}

\item{level}{confidence level for prediction intervals.}

\item{holdout}{whether to use holdout of h observations or not.}
}
\value{
Function returns a model of a class "counter", which contains:
\itemize{
\item model - the name of the constructed model,
\item occurrence - ETS(A,N,N) model for demand intervals (from es() function),
\item sizes - ETS(A,N,N) model for demand sizes (from es() function),
\item fitted - fitted values,
\item forecast - forecasts of the function,
\item lower - lower bound of the prediction interval,
\item upper - upper bound of the prediction interval,
\item actuals - the provided actual values,
\item holdout - the actual values from the holdout (if holdout was set to TRUE),
\item level - confidence level used,
\item probabilty - the probability of success,
\item dispersion - the dispersion variable,
\item alpha - smoothing parameter of the model,
\item initial - the initial value for the time varying mean,
\item accuracy - the error measures for the data if the holdout was TRUE.
}
}
\description{
Construct NegBin model on time series data
}
\details{
The model constructs Negative Binomial model parameterised over the mean and
dispersion parameter using Simple Exponential Smoothing.
}
\examples{
y <- c(rpois(50,0.3),rpois(50,0.8))
test <- negbin(y)

}
\references{
\itemize{
\item Snyder, R. D., Ord, J. K., & Beaumont, A. (2012). Forecasting the
intermittent demand for slow-moving inventories: A modelling approach.
International Journal of Forecasting, 28(2), 485–496.
https://doi.org/10.1016/j.ijforecast.2011.03.009
}
}
\seealso{
\code{\link[counter]{hsp}, \link[smooth]{es}, \link[smooth]{iss}}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.ru}
}
\keyword{models}
\keyword{ts}
