% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{counter}
\alias{counter}
\title{Counts Data Points in Data Frame Group}
\usage{
counter(df, group_var)
}
\arguments{
\item{df}{A dataframe with different groups.}

\item{group_var}{The group of interest.}
}
\value{
A tibble.
}
\description{
Counts Data Points in Data Frame Group
}
\examples{
counter(mtcars, disp )
counter(faithful, eruptions )
counter(faithful, waiting)
}
